tests/light/pyproject.toml
[tool.poetry]
name = "axosyslog-light"
version = "0.2.0"
description = "Lightweight End-to-End Test Framework for AxoSyslog."
authors = ["Andras Mitzki <andras.mitzki@axoflow.com>", "Attila Szakacs <attila.szakacs@axoflow.com>"]
readme = "README.md"
license = "GPL-2.0-or-later"
homepage = "https://github.com/axoflow/axosyslog/tests/light"
repository = "https://github.com/axoflow/axosyslog/tests/light"
keywords = ["axosyslog"]
classifiers = []
packages = [{include = "axosyslog_light", from = "src"}]
include = []
exclude = [
  "src/axosyslog_light/Makefile.am",
]

[tool.poetry.dependencies]
python = "^3.9"
prometheus-client = "^0.21.1"
psutil = "^7.0.0"
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
opentelemetry-sdk = "^1.31.1"
opentelemetry-exporter-otlp-proto-grpc = "^1.31.1"

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.12.1"
isort = "^6.0.1"
pylint = "^3.3.4"
astroid = "^3.3.8"
logilab-common = "^2.0.0"
pre-commit = "^4.1.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
